+++ packages/common/src/compiler/translator.mock.ts
@@ -505,7 +505,7 @@ export const LIGHTDASH_TABLE_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS_SNOWFLAKE: Om
             fieldType: FieldType.DIMENSION,
             description: undefined,
             type: DimensionType.DATE,
-            sql: "DATE_TRUNC('DAY', CONVERT_TIMEZONE('UTC', ${TABLE}.user_created))",
+            sql: "TO_TIMESTAMP_NTZ(DATE_TRUNC('DAY', CONVERT_TIMEZONE('UTC', ${TABLE}.user_created)))",
             name: 'user_created_day',
             table: MODEL_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS.name,
             tableLabel: 'My table',
@@ -523,7 +523,7 @@ export const LIGHTDASH_TABLE_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS_SNOWFLAKE: Om
             fieldType: FieldType.DIMENSION,
             description: undefined,
             type: DimensionType.DATE,
-            sql: "DATE_TRUNC('WEEK', CONVERT_TIMEZONE('UTC', ${TABLE}.user_created))",
+            sql: "TO_TIMESTAMP_NTZ(DATE_TRUNC('WEEK', CONVERT_TIMEZONE('UTC', ${TABLE}.user_created)))",
             name: 'user_created_week',
 
             table: MODEL_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS.name,
@@ -542,7 +542,7 @@ export const LIGHTDASH_TABLE_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS_SNOWFLAKE: Om
             fieldType: FieldType.DIMENSION,
             description: undefined,
             type: DimensionType.DATE,
-            sql: "DATE_TRUNC('MONTH', CONVERT_TIMEZONE('UTC', ${TABLE}.user_created))",
+            sql: "TO_TIMESTAMP_NTZ(DATE_TRUNC('MONTH', CONVERT_TIMEZONE('UTC', ${TABLE}.user_created)))",
             name: 'user_created_month',
 
             table: MODEL_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS.name,
@@ -561,7 +561,7 @@ export const LIGHTDASH_TABLE_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS_SNOWFLAKE: Om
             fieldType: FieldType.DIMENSION,
             description: undefined,
             type: DimensionType.DATE,
-            sql: "DATE_TRUNC('YEAR', CONVERT_TIMEZONE('UTC', ${TABLE}.user_created))",
+            sql: "TO_TIMESTAMP_NTZ(DATE_TRUNC('YEAR', CONVERT_TIMEZONE('UTC', ${TABLE}.user_created)))",
             name: 'user_created_year',
 
             table: MODEL_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS.name,