+++ packages/frontend/src/components/SimpleTable/index.tsx
@@ -17,7 +17,6 @@ const SimpleTable: FC = () => {
         isLoading,
         columnOrder,
         explore,
-
         tableConfig: { itemMap, getHeader, isFilterVisible },
     } = useVisualizationContext();
     const tableItems = resultsData?.rows