+++ packages/frontend/src/components/TableConfigPanel/ColumnConfiguration.tsx
@@ -1,9 +1,11 @@
 import { Button, InputGroup } from '@blueprintjs/core';
+
 import React from 'react';
 import { useVisualizationContext } from '../LightdashVisualization/VisualizationProvider';
 import {
     ColumnConfigurationWrapper,
     ColumnWrapper,
+
 } from './ColumnConfiguration.styles';
 
 export const ColumnConfiguration: React.FC = () => {
@@ -32,6 +34,7 @@ export const ColumnConfiguration: React.FC = () => {
                         />
 
                         <Button
+
                             icon={
                                 isFilterVisible(fieldId)
                                     ? 'eye-off'