+++ CHANGELOG.md
@@ -1,3 +1,59 @@
+## [0.171.4](https://github.com/lightdash/lightdash/compare/0.171.3...0.171.4) (2022-07-05)
+
+
+### Bug Fixes
+
+* timestamp filter ([#2635](https://github.com/lightdash/lightdash/issues/2635)) ([277a904](https://github.com/lightdash/lightdash/commit/277a9043ccdacd632e518cd073f8f70081f9831c))
+
+## [0.171.3](https://github.com/lightdash/lightdash/compare/0.171.2...0.171.3) (2022-07-05)
+
+
+### Bug Fixes
+
+* reset dbt.host_domain field ([#2632](https://github.com/lightdash/lightdash/issues/2632)) ([db22d80](https://github.com/lightdash/lightdash/commit/db22d80453c5dc896c3947cd15c2428cd1b2eef2))
+
+## [0.171.2](https://github.com/lightdash/lightdash/compare/0.171.1...0.171.2) (2022-07-05)
+
+
+### Bug Fixes
+
+* date picker timezone conversion ([#2633](https://github.com/lightdash/lightdash/issues/2633)) ([43661a9](https://github.com/lightdash/lightdash/commit/43661a9bbb6abd975f2d9f6be7d091660ad8a24c))
+
+## [0.171.1](https://github.com/lightdash/lightdash/compare/0.171.0...0.171.1) (2022-07-05)
+
+
+### Bug Fixes
+
+* strip timezone from truncated snowflake fields ([#2629](https://github.com/lightdash/lightdash/issues/2629)) ([f80fa93](https://github.com/lightdash/lightdash/commit/f80fa93dbc24e213bc63da4fb8c3668056bdc21b))
+
+# [0.171.0](https://github.com/lightdash/lightdash/compare/0.170.2...0.171.0) (2022-07-05)
+
+
+### Features
+
+* hide columns table charts ([#2620](https://github.com/lightdash/lightdash/issues/2620)) ([825e327](https://github.com/lightdash/lightdash/commit/825e3272454f040f7240c13f43e32466e8f922cc))
+
+## [0.170.2](https://github.com/lightdash/lightdash/compare/0.170.1...0.170.2) (2022-07-04)
+
+
+### Bug Fixes
+
+* encode uri components in redshift connection string ([#2625](https://github.com/lightdash/lightdash/issues/2625)) ([f25b495](https://github.com/lightdash/lightdash/commit/f25b495906186c00d0dec7423249e0dac3c555b1))
+
+## [0.170.1](https://github.com/lightdash/lightdash/compare/0.170.0...0.170.1) (2022-07-04)
+
+
+### Bug Fixes
+
+* update callout to show dbt 1.1.0 required ([#2626](https://github.com/lightdash/lightdash/issues/2626)) ([1f13cc5](https://github.com/lightdash/lightdash/commit/1f13cc536b535b3f9b0774d19e007aa152bc0a77))
+
+# [0.170.0](https://github.com/lightdash/lightdash/compare/0.169.0...0.170.0) (2022-07-04)
+
+
+### Features
+
+* new table config and show table names property ([#2617](https://github.com/lightdash/lightdash/issues/2617)) ([974b6d6](https://github.com/lightdash/lightdash/commit/974b6d65e578274cdf8c11b20c949451ec6f9d90))
+
 # [0.169.0](https://github.com/lightdash/lightdash/compare/0.168.0...0.169.0) (2022-07-04)